Refactoring

+++ src/main/java/grapher/interactor/shapes/IShape.java
@@ -12,5 +12,5 @@ public interface IShape {
     List<PointData> getAllPointsData();
     ShapeData getData();
     List<PointData> getPointsDataForMorph(int countPoint);
-    int getPerimeter();
+    double getPerimeter();
 }
